refactor: 캐릭터 조회 로직 user 도메인으로 이동
