+++ src/main/java/com/example/inkspire/character/CharacterRepository.java
@@ -2,15 +2,12 @@ package com.example.inkspire.character;
 
 import com.example.inkspire.character.model.Character;
 import java.util.List;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CharacterRepository extends JpaRepository<Character, Long> {
-    List<Character> findAllByUserId(Long userId);
-
     // 캐릭터 별 목표가 생성된 챕터 조회
     @Query(value = "SELECT g.chapter " +
             "FROM characters c " +
