+++ src/main/java/com/example/inkspire/user/UserRepository.java
@@ -1,7 +1,9 @@
 package com.example.inkspire.user;
 
+import com.example.inkspire.character.model.Character;
 import com.example.inkspire.user.model.User;
 import com.example.inkspire.user.model.UserInfoDto;
+import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,6 +12,9 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
+
+    @Query(value = "SELECT * FROM characters c WHERE c.user_id = :userId", nativeQuery = true)
+    List<Character> findCharactersByUserId(Long userId);
 
     @Query(value = "SELECT u.id, u.email, u.nickname, COUNT(s.id) AS endingCount "
             + "FROM user u "
